Move leaf and SCT serialization to a separate file.
This is a better division between RFC 6962 and log specific code.
